추천 서비스 v 0.2.5 TF_IDF 적용으로 검색 결과, 연관 게시물 정확도 향상

+++ src/main/java/io/github/parkjeongwoong/application/blog/dto/ArticleSearchResultDto.java
@@ -18,10 +18,10 @@ public class ArticleSearchResultDto implements Comparable<ArticleSearchResultDto
     private String content;
     private String date;
     private List<Boolean> matchWords = new ArrayList<>();
-    private long priorityScore;
+    private double priorityScore;
     private List<String> searchWords;
 
-    public ArticleSearchResultDto(Article entity, Long priorityScore, List<String> searchWords) {
+    public ArticleSearchResultDto(Article entity, double priorityScore, List<String> searchWords) {
         this.title = entity.getTitle();
         this.category = entity.getCategory();
         this.categoryId = entity.getCategoryId();
